% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombine-package.R
\docType{data}
\name{example_sim_data}
\alias{example_sim_data}
\title{An example dataset of simulation data}
\format{
The format is a list containing the following elements:
\itemize{
\item x: data matrix of the simulation dataset.
\item y: integer vector corresponding to the cluster membership.
}
}
\description{
This dataset has 3 clusters, each of which has 40 data points with 50 informative features and 950 noise features
}
\details{
This dataset can be generated by running:
\code{source(system.file("scripts", "gen_sim_data.R", package = "recombine"))}
followed by
\code{example_sim_data <- gen_sim_data(size_clusters = c(40, 40, 40),
                                       p_inf = 50,
                                       p_noise = 950,
                                       iseed = 1)}
}
\examples{
library(recombine)
example_sim_data

}
\keyword{datasets}
