% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhood_recur_test.R
\name{get_mt_nn}
\alias{get_mt_nn}
\title{get_mt_nn}
\usage{
get_mt_nn(mt_dist, k = 20)
}
\arguments{
\item{mt_dist}{distance matrix of cells. mt_dist needs to be a symmetric matrix or an object of stats::dist.}

\item{k}{number of nearest neighbors. The default is 20.}
}
\value{
\item{indicator matrix of nearest neighbors}{a binary matrix of ncells by ncells, with element \code{mt_nn[i,j]=1} indicating cell j is one of the nearest neighbors of cell i.}
}
\description{
Get indicator matrix of nearest neighbors
}
\details{
Given a distance matrix of cells, get_mt_nn extracts K nearest neighbors of each cell to build an indicator matrix where each row represents a cell's neighborhood status, with 1 indicating a member of nearest neighbors.
}
\examples{
x <- matrix(rnorm(1000), nrow = 100)
mt_nn <- get_mt_nn(dist(x))
mt_nn[1:10, 1:10]

}
