% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhood_recur_test.R
\name{get_rcm_subpop}
\alias{get_rcm_subpop}
\title{get_rcm_subpop}
\usage{
get_rcm_subpop(
  df_rcm_cell,
  df_cell_subpop,
  subpop_name = "cluster",
  nhood_recur_fdr_threshod = 0.05
)
}
\arguments{
\item{df_rcm_cell}{data frame of RCMs per cell, including columns: cell, marker, nhood_zscore, nhood_recur_pval, nhood_recur_fdr.}

\item{df_cell_subpop}{data frame of cell subpopulations, including columns: cell, name of subpopulations (e.g., cluster).}

\item{subpop_name}{name of subpopulations. The default is cluster.}

\item{nhood_recur_fdr_threshod}{threshold for the statistical significance of neighborhood recurrence FDRs. The default is 0.05.}
}
\value{
\item{data frame of RCMs per cell subpopulation}{including columns: subpop, marker, fract_signif_cells, avg_nhood_zscore, PR_AUC, ROC_AUC.}
}
\description{
Recurrent composite markers for cell subpopulations
}
\details{
Given the recurrent composite markers (RCMs) per cell and the membership of cell subpopulations, get_rcm_subpop calculates the average neighborhood Z score and fraction of significant cells for each marker for each cell subpopulation.
}
\examples{
df_rcm_subpop <- get_rcm_subpop(intestine_data$df_rcm_cell, intestine_data$df_cell_subpop)
df_rcm_subpop

}
