% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombine-package.R
\docType{data}
\name{intestine_data}
\alias{intestine_data}
\title{scRNA data and RCMs of mouse intestinal organoids}
\format{
The format is a list containing the following elements:
\itemize{
\item mt_expr: single cell expression data matrix. mt_expr has been normalized and log-transformed.
\item w_nonzero: w values of discriminant markers with nonzero w's.
\item df_cell_subpop: data frame of cell subpopulations.
\item mt_nn: indicator matrix of nearest neighbors. mt_nn is obtained from the nearest neighbors calculated by Seurat based on discriminant markers.
\item df_rcm_cell: data frame of RCMs per cell.
\item df_rcm_subpop: data frame of RCMs per cell subpopulation.
}
}
\description{
The raw single cell expression dataset was obtained from Grün, D., et al. (2015). Single-cell messenger RNA sequencing reveals rare intestinal cell types. Nature, 525(7568), 251–255.
RECOMBINE was applied to this dataset to extract recurrent composite markers (RCMs) for both major and rare cell subpopulations.
}
\examples{
library(recombine)
intestine_data

}
\keyword{datasets}
