% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhood_recur_test.R
\name{nhood_recur_test}
\alias{nhood_recur_test}
\title{nhood_recur_test}
\usage{
nhood_recur_test(mt_expr, mt_nn, verbose = FALSE)
}
\arguments{
\item{mt_expr}{data matrix. Rows are cells and columns are features.}

\item{mt_nn}{indicator matrix of nearest neighbors. mt_nn is a binary matrix of ncells by ncells, with \code{mt_nn[i,j]=1} indicating cell j is one of the nearest neighbors of cell i.}

\item{verbose}{logical variable of progress messages to be printed. The default is FALSE.}
}
\value{
\item{data frame of RCMs per cell}{including columns: cell, marker, nhood_zscore, nhood_recur_pval, nhood_recur_fdr.}
}
\description{
Neighborhood recurrence test
}
\details{
Neighborhood recurrence test determines if markers are recurrently up/down-regulated in the nearest neighborhood of each cell.
}
\examples{
df_rcm_cell <- nhood_recur_test(intestine_data$mt_expr[, names(intestine_data$w_nonzero)], intestine_data$mt_nn)
df_rcm_cell

}
