% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombine-package.R
\docType{package}
\name{recombine-package}
\alias{recombine-package}
\title{recombine package}
\description{
RECOMBINE: Recurrent Composite Markers for Biological Identities with Neighborhood Enrichment
}
\details{
This package includes five novel algorithms for feature selection coupled with clustering. Briefly, we advance SHC (sparse hierarchical clustering) and obtain two algorithms: SHC-SSL (sparse hierarchical clustering with spike-and-slab lasso), which replaces lasso with spike-and-slab lasso to debias the effect of lasso without a priori information, and SHC-FL (sparse hierarchical clustering with fused lasso), which replaces lasso with fused lasso to incorporate prior information of feature order. Further, rSHC, rSHC-SSL, and rSHC-FL, are robust versions of SHC, SHC-SSL, and SHC-FL, respectively, in which selected features are robust to the existence of outliers.
}
\references{
Li X. and Korkut A. (2023) Recurrent composite markers for cell subpopulations with high granularity.
}
\keyword{package}
