% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{recombine_pipeline}
\alias{recombine_pipeline}
\title{recombine_pipeline}
\usage{
recombine_pipeline(
  sobj,
  subpop_name = NULL,
  fixed.hyperparameter = TRUE,
  n_constraint = 50,
  lambda0.fixed = 1,
  lambda0s = seq(0.001, 4000, length = 100),
  lambda1 = 1e-04,
  nperms = 10,
  max_features = 10000,
  pseudocell_embedding_nhvgs = 2000,
  pseudocell_embedding_npcs = 50,
  pseudocell_embedding_batchname = NULL,
  pseudocell_embedding_max.iter.harmony = 100,
  pseudocell_microcluster_res = 50,
  pseudocell_microcluster_res.iter.rate = 0.8,
  allcell_discrim.marker_do.scale = FALSE,
  allcell_discrim.marker_npcs = 20,
  allcell_discrim.marker_neighborhood_k.param = 20,
  nhood_recur_fdr_threshod = 0.05
)
}
\arguments{
\item{sobj}{Seurat object that includes LogNormalized data in layer sobj@assays$RNA$data.}

\item{subpop_name}{name of subpopulations that can be accessed in sobj@meta.data. The default is NULL (no subpopulation markers will be in the output).}

\item{fixed.hyperparameter}{fixed-parameter version (fRECOMBINE).}

\item{n_constraint}{constrained number of discriminant markers.}

\item{lambda0.fixed}{lambda0 value for fRECOMBINE.}

\item{lambda0s}{a sereis of lambda0 values for hyperparameter selection in RECOMBINE.}

\item{lambda1}{lambda1 value.}

\item{nperms}{number of permutation for gap statistic calculation in RECOMBINE.}

\item{max_features}{max number of features before RECOMBINE analysis. If the total feature number exceeds this, HVGs will be selected to match max_features before RECOMBINE analysis.}

\item{pseudocell_embedding_nhvgs}{number of HVGs used in embedding during pseudo-cell construction.}

\item{pseudocell_embedding_npcs}{number of PCs used in embedding during pseudo-cell construction.}

\item{pseudocell_embedding_batchname}{name of batches. If not NULL, it needs to be a string and can be accessed in sobj@meta.data.}

\item{pseudocell_embedding_max.iter.harmony}{max number of iteration for running Harmony to correct batch effect indicated by pseudocell_embedding_batchname.}

\item{pseudocell_microcluster_res}{initial resolution of clustering during pseudo-cell construction.}

\item{pseudocell_microcluster_res.iter.rate}{iterative damping rate of clustering resolution during pseudo-cell construction.}

\item{allcell_discrim.marker_do.scale}{scale during contruction of KNN graph using discriminant markers.}

\item{allcell_discrim.marker_npcs}{number of PCs during contruction of KNN graph using discriminant markers.}

\item{allcell_discrim.marker_neighborhood_k.param}{K of the KNN graph using discriminant markers.}

\item{nhood_recur_fdr_threshod}{threshold for the statistical significance of neighborhood recurrence FDRs. The default is 0.05.}
}
\value{
\item{list of data frames}{including features with nonzero weights, recurrent composite markers of subpopulations, recurrent composite markers of individual cells, and gap statistic output from a series of lambda0s.}
}
\description{
Streamlined RECOMBNE pipeline
}
\details{
Given a Seurat object with log-normalized data and optionally subpopulation annotation, this pipeline output discriminant markers of the dataset and recurrent composite markers for subpopulations and individual cells.
}
\examples{
recombine.out <- recombine_pipeline(sobj, subpop_name = "celltype")

}
