% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fl.R
\name{SHC_FL}
\alias{SHC_FL}
\title{SHC_FL}
\usage{
SHC_FL(
  x,
  lambda1,
  lambda2,
  standardize.arrays = FALSE,
  dissimilarity = c("squared.distance", "absolute.value"),
  method = c("average", "complete", "single", "centroid", "ward.D", "ward.D2",
    "mcquitty", "median"),
  max.iter = 100,
  iseed = NULL,
  silent = FALSE
)
}
\arguments{
\item{x}{data matrix. Rows are samples, while columns are features.}

\item{lambda1}{sparsity parameter.}

\item{lambda2}{successive difference sparsity parameter.}

\item{standardize.arrays}{should the data matrix be standardized? The default is FALSE.}

\item{dissimilarity}{type of dissimilarity metric. This should be either "squared.distance" or "absolute.value". The default is "squared.distance".}

\item{method}{the agglomeration method to be used. This should be (an unambiguous abbreviation of) one of "ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC). The default is "average".}

\item{max.iter}{maximum number of iterations. The default is 500.}

\item{iseed}{seed of random numbers used for initializing w. If Null, all elements of w are initialized as a fixed value of 1.0/sqrt(p). Otherwise, the initial w elements are uniformly distributed in \code{[0, 2.0/sqrt(p)]}. The default is NULL.}

\item{silent}{should progress messages be suppressed? The default is FALSE.}
}
\value{
\item{lassotype}{abbreviated name of lasso type.}
\item{lassotype_full}{full name of lasso type.}
\item{u}{the optimal weighted dissimilarity matrix.}
\item{w}{the optimal weights.}
\item{crit}{the optimal objective.}
\item{iter}{number of iteration steps.}
\item{flsa_converged}{indicator of convergence of the FLSA (Fused Lasso Signal Approximator) solver.}
\item{w_l0norm}{L0 norm of w.}
\item{w_diff_l0norm}{L0 norm of successive differences of w's.}
\item{nonzero_w_indices}{indices of nonzero w's.}
\item{hyp1}{hyperparameter lambda1 used.}
\item{hyp2}{hyperparameter lambda2 used.}
\item{hc}{an object of class hclust by running hierarchical clustering on u.}
}
\description{
Sparse hierarchical clustering with fused lasso
}
\details{
SHC_FL employs a sparse fused lasso penalty to perform feature selection coupled with clustering.
It encourages sparsity of features, and encourages neighboring features to be similar and some to be identical.
}
\examples{
d <- example_sim_data

# reorder features based on generative process
order <- c()
for (i in 1:n_clusters) {
  order <- c(order, seq(from = i, to = p_inf, by = n_clusters))
}
order <- c(order, (p_inf+1):p)

mt <- d$x[, order]

# run
result <- SHC_FL(d$x,
                 lambda1 = 200,
                 lambda2 = 3000)
result

}
