% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_lagrange.R
\name{SHC_lagrange}
\alias{SHC_lagrange}
\title{SHC_lagrange}
\usage{
SHC_lagrange(
  x,
  lambda,
  standardize.arrays = FALSE,
  dissimilarity = c("squared.distance", "absolute.value"),
  method = c("average", "complete", "single", "centroid", "ward.D", "ward.D2",
    "mcquitty", "median"),
  max.iter = 100,
  iseed = NULL,
  silent = FALSE
)
}
\arguments{
\item{x}{data matrix. Rows are samples, while columns are features.}

\item{lambda}{lasso variance parameter.}

\item{standardize.arrays}{should the data matrix be standardized? The default is FALSE.}

\item{dissimilarity}{type of dissimilarity metric. This should be either "squared.distance" or "absolute.value". The default is "squared.distance".}

\item{method}{the agglomeration method to be used. This should be (an unambiguous abbreviation of) one of "ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC). The default is "average".}

\item{max.iter}{maximum number of iterations. The default is 500.}

\item{iseed}{seed of random numbers used for initializing w. If Null, all elements of w are initialized as a fixed value of 1.0/sqrt(p). Otherwise, the initial w elements are uniformly distributed in \code{[0, 2.0/sqrt(p)]}. The default is NULL.}

\item{silent}{should progress messages be suppressed? The default is FALSE.}
}
\value{
\item{lassotype}{abbreviated name of lasso type.}
\item{lassotype_full}{full name of lasso type.}
\item{u}{the optimal weighted dissimilarity matrix.}
\item{w}{the optimal weights.}
\item{crit}{the optimal objective.}
\item{iter}{number of iteration steps.}
\item{w_l0norm}{L0 norm of w.}
\item{nonzero_w_indices}{indices of nonzero w's.}
\item{hyp}{hyperparameter used (i.e., lambda).}
\item{hc}{an object of class hclust by running hierarchical clustering on u.}
}
\description{
Sparse hierarchical clustering via a Lagrange form
}
\details{
SHC_lagrange employs a lasso penalty to perform feature selection coupled with clustering.
In the objective, lasso penalty is written in a Lagrange form rather than as a constraint.
}
\examples{
d <- example_sim_data
result <- SHC_lagrange(d$x, lambda = 300)
result

}
