% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssl.R
\name{SHC_SSL}
\alias{SHC_SSL}
\title{SHC_SSL}
\usage{
SHC_SSL(
  x,
  penalty = c("adaptive", "separable"),
  lambda0,
  lambda1,
  theta = 0.5,
  a = 1,
  b,
  eps = 0.001,
  standardize.arrays = FALSE,
  dissimilarity = c("squared.distance", "absolute.value"),
  method = c("average", "complete", "single", "centroid", "ward.D", "ward.D2",
    "mcquitty", "median"),
  max.iter = 500,
  iseed = NULL,
  silent = FALSE
)
}
\arguments{
\item{x}{data matrix. Rows are samples, while columns are features.}

\item{penalty}{type of spike-and-slab lasso penalty. This should be either "adaptive" (with a random theta, where theta ~ B(a,b)) or "separable" (with a fixed theta). The default is "adaptive".}

\item{lambda0}{spike variance parameter.}

\item{lambda1}{slab variance parameter. Note lambda1 should be no greater than lambda0.}

\item{theta}{prior mixing proportion. For "separable" penalty, this value is fixed. For "adaptive" penalty, this value is used as a starting value. The default is 0.5.}

\item{a}{parameter of the beta prior B(a,b) for the adaptive penalty. The default is 1.}

\item{b}{parameter of the beta prior B(a,b) for the adaptive penalty. The default is \code{ncol(x)}.}

\item{eps}{convergence criterion. Converged when all component-wise differences in w are less than eps. The default is 0.001.}

\item{standardize.arrays}{should the data matrix be standardized? The default is FALSE.}

\item{dissimilarity}{type of dissimilarity metric. This should be either "squared.distance" or "absolute.value". The default is "squared.distance".}

\item{method}{the agglomeration method to be used. This should be (an unambiguous abbreviation of) one of "ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC). The default is "average".}

\item{max.iter}{maximum number of iterations. The default is 500.}

\item{iseed}{seed of random numbers used for initializing w. If Null, all elements of w are initialized as a fixed value of 1.0/sqrt(p). Otherwise, the initial w elements are uniformly distributed in \code{[0, 2.0/sqrt(p)]}. The default is NULL.}

\item{silent}{should progress messages be suppressed? The default is FALSE.}
}
\value{
\item{lassotype}{abbreviated name of lasso type.}
\item{lassotype_full}{full name of lasso type.}
\item{u}{the optimal weighted dissimilarity matrix.}
\item{w}{the optimal weights.}
\item{crit}{the optimal objective.}
\item{iter}{number of iteration steps.}
\item{w_l0norm}{L0 norm of w.}
\item{nonzero_w_indices}{indices of nonzero w's.}
\item{hyp}{hyperparameter used (i.e., lambda0).}
\item{hc}{an object of class hclust by running hierarchical clustering on u.}
}
\description{
Sparse hierarchical clustering with spike-and-slab lasso
}
\details{
SHC_SSL employs a spike-and-slab lasso penalty to approach the theoretically ideal feature selection coupled with clustering.
}
\examples{
d <- example_sim_data
result <- SHC_SSL(d$x,
                  lambda0 = 200,
                  lambda1 = 0.001)
result

}
